requirements.txt
requirements.txt
swarms>=0.1.0
openai>=1.0.0
anthropic>=0.5.0
google-genai>=0.3.0
chromadb>=0.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
docker>=6.0.0
requests>=2.0.0
numpy>=1.0.0
pandas>=2.0.0
scikit-learn>=1.0.0
tqdm>=4.0.0
sentence-transformers>=2.2.0
